+++ CHANGES.rst
@@ -5,7 +5,14 @@ Changelog
 1.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Make the promotions to be optional to be rendered with results
+  [frapell]
+
+- Allow to customize the arguments sent to cse with adapters
+  [frapell]
+
+- When a request to Google fails, retry up to 5 times
+  [frapell]
 
 
 1.0.0 (2018-01-15)